Solved the refresh page issue
Resolved #1.  The issue was with the getting an access token from the "code"
returned from the OAuth dialog.  If we refreshed the page too quickly, the
code would "be in use".  If the page went stale, the code had expired.

By saving the short-term access token we can still access the user data and
maintain if a user is logged in.

We violate DRY a bit here, so the code will need to be refactored a bit, as
per #8.