Implemented HTTP authentication to the API endpoints.
The API endpoints are now protected with HTTP basic authentication.  This
should be alright, because the connection is encrypted. At the moment the
HTTP basic authentication user and password combination is hardcoded in the
views.py. This should be moved the the database #11.

Unit tests were implemented to test the API and the authentication associated
with it.  The API could do with some more test coverage #13.

Added requests and Flask-HTTPAuth to the package requirements.